requirements.txt
python-telegram-bot==22.5
python-dotenv>=1.0.0
httpx>=0.24.0
google-auth>=2.20.0
openai-whisper>=20231117
pydub>=0.25.1
ffmpeg-python>=0.2.0
gTTS>=2.3.2
requests>=2.31.0
pytz>=2024.1
streamlit>=1.28.0
